Add files via upload
GUI.py: This is the interface from which one can detect the facial expression of uploaded image. This the final step after successfully training the ML model.
Kaggle_emotion_detection.ipynb: This is the python notebook for training the machine learning model.
model.weights.h5: This is file containing model weights. Model weights are a subset of a model's parameters that determine how input tokens are transformed as they pass through the model's layers.
haarcascade-frontalface_default.xml: This is a cascade classifier file which identify the presence of faces in pictures.